JitArm64: Keep float constants in fixed registers
PR 9768 was an attempt to make the register cache keep track of float
constants so that we would only have to emit them once per block instead
of once per use. However, the code of that PR was a bit complicated.

This PR offers a simpler solution: Since there are only three constants
we want to keep in registers, simply reserve three registers for them.
This has the advantage of letting us keep the constants in registers
across blocks, decreasing code size even further. The downside is that
fewer registers are available for the register cache, but with how many
registers AArch64 has, I think it's a good tradeoff.