Jit64: Disable branch merging for SO bit
Dolphin's JITs don't accurately emulate the SO bit of condition
registers. However, exactly which inaccurate behavior they end up using
depends on whether branch merging is applied. This is in theory a
problem for cross-determinism between Jit64 and JitArm64, because
JitArm64 doesn't support branch merging. But since games generally don't
branch based on the SO bit, I'm not sure if it's a problem in practice.

This commit disables branch merging for the SO bit so that we get the
same behavior in Jit64 as in JitArm64.